feat: more perfy
